requirements.txt
streamlit>=1.28.0
opencv-python>=4.5.0
numpy>=1.19.0
pytesseract>=0.3.8
Pillow>=8.0.0
pdf2image>=1.16.0
scikit-learn>=0.24.0
nltk>=3.6.0
python-Levenshtein>=0.12.0
transformers>=4.0.0
torch>=1.8.0
spacy>=3.0.0
textblob>=0.15.3
fuzzywuzzy>=0.18.0
python-docx>=0.8.11
PyPDF2>=2.0.0
fpdf2>=2.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
